requirements.txt
fastapi
uvicorn
langchain_community
faiss-cpu
sentence-transformers
llama-cpp-python
pypdf
beautifulsoup4
requests
